gio_win32::prelude

Trait FilterOutputStreamExt

pub trait FilterOutputStreamExt: IsA<FilterOutputStream> + 'static {
    // Provided methods
    fn base_stream(&self) -> OutputStream { ... }
    fn closes_base_stream(&self) -> bool { ... }
    fn set_close_base_stream(&self, close_base: bool) { ... }
    fn connect_close_base_stream_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
}

Trait containing all FilterOutputStream methods.

Implementors

BufferedOutputStream, ConverterOutputStream, DataOutputStream, FilterOutputStream

Provided Methods

fn base_stream(&self) -> OutputStream

Gets the base stream for the filter stream.

Returns

a OutputStream.

fn closes_base_stream(&self) -> bool

Returns whether the base stream will be closed when @self is closed.

Returns

TRUE if the base stream will be closed.

fn set_close_base_stream(&self, close_base: bool)

Sets whether the base stream will be closed when @self is closed.

close_base

TRUE to close the base stream.

fn connect_close_base_stream_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.