gio_win32::prelude

Trait SocketControlMessageExtManual

pub trait SocketControlMessageExtManual: Sized + IsA<SocketControlMessage> {
    // Provided method
    fn serialize(&self, data: &mut [u8]) { ... }
}

Provided Methods

fn serialize(&self, data: &mut [u8])

Converts the data in the message to bytes placed in the message.

@data is guaranteed to have enough space to fit the size returned by g_socket_control_message_get_size() on this object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.