gio_win32::prelude

Trait TlsConnectionExt

pub trait TlsConnectionExt: IsA<TlsConnection> + 'static {

    // Provided methods
    fn emit_accept_certificate(
        &self,
        peer_cert: &impl IsA<TlsCertificate>,
        errors: TlsCertificateFlags,
    ) -> bool { ... }
    fn certificate(&self) -> Option<TlsCertificate> { ... }
    fn ciphersuite_name(&self) -> Option<GString> { ... }
    fn database(&self) -> Option<TlsDatabase> { ... }
    fn interaction(&self) -> Option<TlsInteraction> { ... }
    fn negotiated_protocol(&self) -> Option<GString> { ... }
    fn peer_certificate(&self) -> Option<TlsCertificate> { ... }
    fn peer_certificate_errors(&self) -> TlsCertificateFlags { ... }
    fn protocol_version(&self) -> TlsProtocolVersion { ... }
    fn rehandshake_mode(&self) -> TlsRehandshakeMode { ... }
    fn requires_close_notify(&self) -> bool { ... }
    fn handshake(
        &self,
        cancellable: Option<&impl IsA<Cancellable>>,
    ) -> Result<(), Error> { ... }
    fn handshake_async<P>(
        &self,
        io_priority: Priority,
        cancellable: Option<&impl IsA<Cancellable>>,
        callback: P,
    )
       where P: FnOnce(Result<(), Error>) + 'static { ... }
    fn handshake_future(
        &self,
        io_priority: Priority,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>>>> { ... }
    fn set_certificate(&self, certificate: &impl IsA<TlsCertificate>) { ... }
    fn set_database(&self, database: Option<&impl IsA<TlsDatabase>>) { ... }
    fn set_interaction(&self, interaction: Option<&impl IsA<TlsInteraction>>) { ... }
    fn set_rehandshake_mode(&self, mode: TlsRehandshakeMode) { ... }
    fn set_require_close_notify(&self, require_close_notify: bool) { ... }
    fn advertised_protocols(&self) -> Vec<GString> { ... }
    fn base_io_stream(&self) -> Option<IOStream> { ... }
    fn connect_accept_certificate<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self, &TlsCertificate, TlsCertificateFlags) -> bool + 'static { ... }
    fn connect_advertised_protocols_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_certificate_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_ciphersuite_name_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_database_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_interaction_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_negotiated_protocol_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_peer_certificate_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_peer_certificate_errors_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_protocol_version_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_rehandshake_mode_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }
    fn connect_require_close_notify_notify<F>(&self, f: F) -> SignalHandlerId
       where F: Fn(&Self) + 'static { ... }

}

Trait containing all TlsConnection methods.

Implementors

TlsClientConnection, TlsConnection, TlsServerConnection

Provided Methods

fn emit_accept_certificate( &self, peer_cert: &impl IsA<TlsCertificate>, errors: TlsCertificateFlags, ) -> bool

Used by #GTlsConnection implementations to emit the #GTlsConnection::accept-certificate signal.

peer_cert

the peer’s #GTlsCertificate

errors

the problems with @peer_cert

Returns

true if one of the signal handlers has returned true to accept @peer_cert

fn certificate(&self) -> Option<TlsCertificate>

Gets @self’s certificate, as set by g_tls_connection_set_certificate().

Returns

@self’s certificate, or None

fn ciphersuite_name(&self) -> Option<GString>

Available on crate feature v2_70 only.

Returns the name of the current TLS ciphersuite, or None if the connection has not handshaked or has been closed. Beware that the TLS backend may use any of multiple different naming conventions, because OpenSSL and GnuTLS have their own ciphersuite naming conventions that are different from each other and different from the standard, IANA- registered ciphersuite names. The ciphersuite name is intended to be displayed to the user for informative purposes only, and parsing it is not recommended.

Returns

The name of the current TLS ciphersuite, or None

fn database(&self) -> Option<TlsDatabase>

Gets the certificate database that @self uses to verify peer certificates. See g_tls_connection_set_database().

Returns

the certificate database that @self uses or None

fn interaction(&self) -> Option<TlsInteraction>

Get the object that will be used to interact with the user. It will be used for things like prompting the user for passwords. If None is returned, then no user interaction will occur for this connection.

Returns

The interaction object.

fn negotiated_protocol(&self) -> Option<GString>

Available on crate feature v2_60 only.

Gets the name of the application-layer protocol negotiated during the handshake.

If the peer did not use the ALPN extension, or did not advertise a protocol that matched one of @self’s protocols, or the TLS backend does not support ALPN, then this will be None. See g_tls_connection_set_advertised_protocols().

Returns

the negotiated protocol, or None

fn peer_certificate(&self) -> Option<TlsCertificate>

Gets @self’s peer’s certificate after the handshake has completed or failed. (It is not set during the emission of #GTlsConnection::accept-certificate.)

Returns

@self’s peer’s certificate, or None

fn peer_certificate_errors(&self) -> TlsCertificateFlags

Gets the errors associated with validating @self’s peer’s certificate, after the handshake has completed or failed. (It is not set during the emission of #GTlsConnection::accept-certificate.)

See #GTlsConnection:peer-certificate-errors for more information.

Returns

@self’s peer’s certificate errors

fn protocol_version(&self) -> TlsProtocolVersion

Available on crate feature v2_70 only.

Returns the current TLS protocol version, which may be TlsProtocolVersion::Unknown if the connection has not handshaked, or has been closed, or if the TLS backend has implemented a protocol version that is not a recognized #GTlsProtocolVersion.

Returns

The current TLS protocol version

fn rehandshake_mode(&self) -> TlsRehandshakeMode

👎Deprecated: Since 2.60

Gets @self rehandshaking mode. See g_tls_connection_set_rehandshake_mode() for details.

Deprecated since 2.60

Changing the rehandshake mode is no longer required for compatibility. Also, rehandshaking has been removed from the TLS protocol in TLS 1.3.

Returns

TlsRehandshakeMode::Safely

fn requires_close_notify(&self) -> bool

Tests whether or not @self expects a proper TLS close notification when the connection is closed. See g_tls_connection_set_require_close_notify() for details.

Returns

true if @self requires a proper TLS close notification.

fn handshake( &self, cancellable: Option<&impl IsA<Cancellable>>, ) -> Result<(), Error>

Attempts a TLS handshake on @self.

On the client side, it is never necessary to call this method; although the connection needs to perform a handshake after connecting (or after sending a “STARTTLS”-type command), #GTlsConnection will handle this for you automatically when you try to send or receive data on the connection. You can call g_tls_connection_handshake() manually if you want to know whether the initial handshake succeeded or failed (as opposed to just immediately trying to use @self to read or write, in which case, if it fails, it may not be possible to tell if it failed before or after completing the handshake), but beware that servers may reject client authentication after the handshake has completed, so a successful handshake does not indicate the connection will be usable.

Likewise, on the server side, although a handshake is necessary at the beginning of the communication, you do not need to call this function explicitly unless you want clearer error reporting.

Previously, calling g_tls_connection_handshake() after the initial handshake would trigger a rehandshake; however, this usage was deprecated in GLib 2.60 because rehandshaking was removed from the TLS protocol in TLS 1.3. Since GLib 2.64, calling this function after the initial handshake will no longer do anything.

When using a #GTlsConnection created by #GSocketClient, the #GSocketClient performs the initial handshake, so calling this function manually is not recommended.

#GTlsConnection::accept_certificate may be emitted during the handshake.

cancellable

a #GCancellable, or None

Returns

success or failure

fn handshake_async<P>( &self, io_priority: Priority, cancellable: Option<&impl IsA<Cancellable>>, callback: P, )
where P: FnOnce(Result<(), Error>) + 'static,

Asynchronously performs a TLS handshake on @self. See g_tls_connection_handshake() for more information.

io_priority

the I/O priority of the request

cancellable

a #GCancellable, or None

callback

callback to call when the handshake is complete

fn handshake_future( &self, io_priority: Priority, ) -> Pin<Box<dyn Future<Output = Result<(), Error>>>>

fn set_certificate(&self, certificate: &impl IsA<TlsCertificate>)

This sets the certificate that @self will present to its peer during the TLS handshake. For a #GTlsServerConnection, it is mandatory to set this, and that will normally be done at construct time.

For a #GTlsClientConnection, this is optional. If a handshake fails with TlsError::CertificateRequired, that means that the server requires a certificate, and if you try connecting again, you should call this method first. You can call g_tls_client_connection_get_accepted_cas() on the failed connection to get a list of Certificate Authorities that the server will accept certificates from.

(It is also possible that a server will allow the connection with or without a certificate; in that case, if you don’t provide a certificate, you can tell that the server requested one by the fact that g_tls_client_connection_get_accepted_cas() will return non-None.)

certificate

the certificate to use for @self

fn set_database(&self, database: Option<&impl IsA<TlsDatabase>>)

Sets the certificate database that is used to verify peer certificates. This is set to the default database by default. See g_tls_backend_get_default_database(). If set to None, then peer certificate validation will always set the TlsCertificateFlags::UNKNOWN_CA error (meaning #GTlsConnection::accept-certificate will always be emitted on client-side connections, unless that bit is not set in #GTlsClientConnection:validation-flags).

There are nonintuitive security implications when using a non-default database. See #GTlsConnection:database for details.

database

a #GTlsDatabase

fn set_interaction(&self, interaction: Option<&impl IsA<TlsInteraction>>)

Set the object that will be used to interact with the user. It will be used for things like prompting the user for passwords.

The @interaction argument will normally be a derived subclass of #GTlsInteraction. None can also be provided if no user interaction should occur for this connection.

interaction

an interaction object, or None

fn set_rehandshake_mode(&self, mode: TlsRehandshakeMode)

👎Deprecated: Since 2.60

Since GLib 2.64, changing the rehandshake mode is no longer supported and will have no effect. With TLS 1.3, rehandshaking has been removed from the TLS protocol, replaced by separate post-handshake authentication and rekey operations.

Deprecated since 2.60

Changing the rehandshake mode is no longer required for compatibility. Also, rehandshaking has been removed from the TLS protocol in TLS 1.3.

mode

the rehandshaking mode

fn set_require_close_notify(&self, require_close_notify: bool)

Sets whether or not @self expects a proper TLS close notification before the connection is closed. If this is true (the default), then @self will expect to receive a TLS close notification from its peer before the connection is closed, and will return a TlsError::Eof error if the connection is closed without proper notification (since this may indicate a network error, or man-in-the-middle attack).

In some protocols, the application will know whether or not the connection was closed cleanly based on application-level data (because the application-level data includes a length field, or is somehow self-delimiting); in this case, the close notify is redundant and sometimes omitted. (TLS 1.1 explicitly allows this; in TLS 1.0 it is technically an error, but often done anyway.) You can use g_tls_connection_set_require_close_notify() to tell @self to allow an “unannounced” connection close, in which case the close will show up as a 0-length read, as in a non-TLS #GSocketConnection, and it is up to the application to check that the data has been fully received.

Note that this only affects the behavior when the peer closes the connection; when the application calls g_io_stream_close() itself on @self, this will send a close notification regardless of the setting of this property. If you explicitly want to do an unclean close, you can close @self’s #GTlsConnection:base-io-stream rather than closing @self itself, but note that this may only be done when no other operations are pending on @self or the base I/O stream.

require_close_notify

whether or not to require close notification

fn advertised_protocols(&self) -> Vec<GString>

Available on crate feature v2_60 only.

The list of application-layer protocols that the connection advertises that it is willing to speak. See g_tls_connection_set_advertised_protocols().

fn base_io_stream(&self) -> Option<IOStream>

The #GIOStream that the connection wraps. The connection holds a reference to this stream, and may run operations on the stream from other threads throughout its lifetime. Consequently, after the #GIOStream has been constructed, application code may only run its own operations on this stream when no #GIOStream operations are running.

fn connect_accept_certificate<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self, &TlsCertificate, TlsCertificateFlags) -> bool + 'static,

Emitted during the TLS handshake after the peer certificate has been received. You can examine @peer_cert’s certification path by calling g_tls_certificate_get_issuer() on it.

For a client-side connection, @peer_cert is the server’s certificate, and the signal will only be emitted if the certificate was not acceptable according to @conn’s #GTlsClientConnection:validation_flags. If you would like the certificate to be accepted despite @errors, return true from the signal handler. Otherwise, if no handler accepts the certificate, the handshake will fail with TlsError::BadCertificate.

GLib guarantees that if certificate verification fails, this signal will be emitted with at least one error will be set in @errors, but it does not guarantee that all possible errors will be set. Accordingly, you may not safely decide to ignore any particular type of error. For example, it would be incorrect to ignore TlsCertificateFlags::EXPIRED if you want to allow expired certificates, because this could potentially be the only error flag set even if other problems exist with the certificate.

For a server-side connection, @peer_cert is the certificate presented by the client, if this was requested via the server’s #GTlsServerConnection:authentication_mode. On the server side, the signal is always emitted when the client presents a certificate, and the certificate will only be accepted if a handler returns true.

Note that if this signal is emitted as part of asynchronous I/O in the main thread, then you should not attempt to interact with the user before returning from the signal handler. If you want to let the user decide whether or not to accept the certificate, you would have to return false from the signal handler on the first attempt, and then after the connection attempt returns a TlsError::BadCertificate, you can interact with the user, and if the user decides to accept the certificate, remember that fact, create a new connection, and return true from the signal handler the next time.

If you are doing I/O in another thread, you do not need to worry about this, and can simply block in the signal handler until the UI thread returns an answer.

peer_cert

the peer’s #GTlsCertificate

errors

the problems with @peer_cert.

Returns

true to accept @peer_cert (which will also immediately end the signal emission). false to allow the signal emission to continue, which will cause the handshake to fail if no one else overrides it.

fn connect_advertised_protocols_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

Available on crate feature v2_60 only.

fn connect_certificate_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

fn connect_ciphersuite_name_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

Available on crate feature v2_70 only.

fn connect_database_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

fn connect_interaction_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

fn connect_negotiated_protocol_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

Available on crate feature v2_60 only.

fn connect_peer_certificate_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

fn connect_peer_certificate_errors_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

fn connect_protocol_version_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

Available on crate feature v2_70 only.

fn connect_rehandshake_mode_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

👎Deprecated: Since 2.60

fn connect_require_close_notify_notify<F>(&self, f: F) -> SignalHandlerId
where F: Fn(&Self) + 'static,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.