gio_win32::prelude

Trait Win32OutputStreamExt

pub trait Win32OutputStreamExt: IsA<OutputStream> + 'static {
    // Provided methods
    fn closes_handle(&self) -> bool { ... }
    unsafe fn set_close_handle(&self, close_handle: bool) { ... }
    fn connect_close_handle_notify<F: Fn(&Self) + 'static>(
        &self,
        f: F,
    ) -> SignalHandlerId { ... }
}

Trait containing all OutputStream methods.

Implementors

OutputStream

Provided Methods

fn closes_handle(&self) -> bool

Returns whether the handle of @self will be closed when the stream is closed.

Returns

true if the handle is closed when done

unsafe fn set_close_handle(&self, close_handle: bool)

Sets whether the handle of @self shall be closed when the stream is closed.

close_handle

true to close the handle when done

fn connect_close_handle_notify<F: Fn(&Self) + 'static>( &self, f: F, ) -> SignalHandlerId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.